Added --no-internet and --load-image options

+++ ecs-single-node/step1_ecs_singlenode_install.py
@@ -432,7 +432,7 @@ def cmdline(command):
     return process.communicate()[0]
 
 
-def modify_container_conf_func():
+def modify_container_conf_func(no_internet):
     try:
         logger.info("Backup object properties files")
         os.system(
@@ -495,19 +495,21 @@ def modify_container_conf_func():
         os.system(
             "docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone cp /host/ssm.object.properties /opt/storageos/conf/ssm.object.properties")
 
-        logger.info("Adding python setuptools to container")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone wget https://bootstrap.pypa.io/ez_setup.py")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone python ez_setup.py")
 
-        logger.info("Adding python requests library to container")
-        os.system(
-            "docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone curl -OL https://github.com/kennethreitz/requests/tarball/master")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone tar zxvf master -C /tmp")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t -i ecsstandalone bash -c \"cd /tmp/kennethreitz-requests-* && python setup.py install\"")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone wget https://bootstrap.pypa.io/ez_setup.py")
-        logger.info("Cleaning up python packages")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone rm master")
-        os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone rm setuptools-20.0.zip")
+        if not no_internet:
+            logger.info("Adding python setuptools to container")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone wget https://bootstrap.pypa.io/ez_setup.py")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone python ez_setup.py")
+    
+            logger.info("Adding python requests library to container")
+            os.system(
+                "docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone curl -OL https://github.com/kennethreitz/requests/tarball/master")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone tar zxvf master -C /tmp")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t -i ecsstandalone bash -c \"cd /tmp/kennethreitz-requests-* && python setup.py install\"")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone wget https://bootstrap.pypa.io/ez_setup.py")
+            logger.info("Cleaning up python packages")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone rm master")
+            os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t  ecsstandalone rm setuptools-20.0.zip")
 
         logger.info("Flush VNeST data")
         os.system("docker "+' '.join(DockerCommandLineFlags)+" exec -t ecsstandalone rm -rf /data/vnest/vnest-main/*")
@@ -550,6 +552,19 @@ def getAuthToken(ECSNode, User, Password):
             # logger.info("Attempting to authenticate for {} minutes.".format(i%2))
 
     logger.fatal("Authentication service not yet started.")
+
+def docker_load_image(imagefile):
+    """
+    Loads the specified docker image file.
+    """
+    try:
+        logger.info("Loading docker image file %s" % imagefile)
+        res = subprocess.check_output("docker load -i \"%s\"" % imagefile, shell=True)
+    except Exception as ex:
+        logger.exception(ex)
+        logger.fatal("Error loading docker image file %s" % imagefile)
+        sys.exit(13)
+
 
 
 def docker_cleanup_old_images():
@@ -667,10 +682,18 @@ def main():
     parser.add_argument('--use-urandom', dest='use_urandom', action='store_true', default=False,
                         help='If present, /dev/random will be mapped to /dev/urandom on the host.  If you container starts up slow the first time could help.',
                         required=False)
+    parser.add_argument('--no-internet', dest='no_internet', action='store_true', default=False,
+                        help='When specified, do not perform any actions that require an Internet connection.',
+                        required=False)
+    parser.add_argument('--load-image', dest='image_file', nargs='?',
+                        help='If present, gives the name of a docker image file to load.',
+                        required=False)
+
     parser.set_defaults(container_config=False)
     parser.set_defaults(cleanup=False)
     parser.set_defaults(imagename="emccorp/ecs-software-2.2")
     parser.set_defaults(imagetag="latest")
+    parser.set_defaults(image_file=False)
     args = parser.parse_args()
 
 
@@ -686,6 +709,16 @@ def main():
     print("Disk[s]: %s" % args.disks)
     print("Docker Image Name: %s" % args.imagename)
     print("Docker Image Tag: %s" % args.imagetag)
+    if(args.image_file):
+        print("Docker Image File: %s" % args.image_file)
+
+    print("Use Internet to download image and packages: %s" % (not args.no_internet))
+
+    # If loading an image, make sure it exists.
+    if args.image_file and not os.path.exists(args.image_file):
+        logger.error("The specified docker image file %s does not exist." % args.image_file)
+        sys.exit(12)
+
 
     # Check if only wants to run the Container Configuration section
     if args.container_config:
@@ -730,11 +763,15 @@ def main():
 
     logger.info("Starting Step 1: Configuration of Host Machine to run the ECS Docker Container: {}".format(docker_image_name))
 
-    yum_func()
-    package_install_func()
+    if not args.no_internet:
+        yum_func()
+        package_install_func()
     update_selinux_os_configuration()
     prep_file_func()
-    docker_pull_func(docker_image_name)
+    if args.image_file:
+        docker_load_image(args.image_file)
+    if not args.no_internet:
+        docker_pull_func(docker_image_name)
     hosts_file_func(args.hostname, ethernet_adapter_name)
     network_file_func(ethernet_adapter_name)
     seeds_file_func(ethernet_adapter_name)
@@ -743,7 +780,7 @@ def main():
     directory_files_conf_func()
     set_docker_configuration_func()
     execute_docker_func(docker_image_name, args.use_urandom)
-    modify_container_conf_func()
+    modify_container_conf_func(args.no_internet)
     getAuthToken(ip_address,"root","ChangeMe")
     logger.info(
         "Step 1 Completed.  Navigate to the administrator website that is available from any of the ECS data nodes. \
